Update Sanity API configuration and fetch functions
